+++ CHANGELOG.md
@@ -16,10 +16,20 @@ Types of changes:
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-04-03
+
+### Added
+- Add new Network, Access, and Kubernetes cluster endpoints
+
+### Changed
+- Move base resource to its own file/class
+- Cleaned up a few of the classes
+
 ## [0.1.0] - 2020-07-01
 
 ### Added
 - Initial import of ecHome python SDK
 
-[unreleased]: https://github.com/mgtrrz/echome-python-sdk/compare/0.1.0...HEAD
+[unreleased]: https://github.com/mgtrrz/echome-python-sdk/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/mgtrrz/echome-python-sdk/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/mgtrrz/echome-python-sdk/releases/tag/0.1.0